backend/requirements.txt
# Python deps — install with: pip install -r requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.35
alembic==1.13.2

# MCP (official Python SDK by Anthropic)
mcp==1.3.0

# LLM
groq==0.11.0              # Groq — free tier, ultra fast, Llama 3.3

# Auth
python-jose[cryptography]==3.3.0   # JWT tokens
passlib[bcrypt]==1.7.4             # password hashing
bcrypt==3.2.2                      # pinned — passlib 1.7.4 incompatible with bcrypt 4+

# External APIs
google-api-python-client==2.143.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
httpx==0.27.2
resend==2.4.0             # email — 3000 free/month, simpler than SendGrid
slack-sdk==3.31.0

# Utilities
python-dotenv==1.0.1
email-validator==2.3.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
